backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
dj-database-url==2.1.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
argon2-cffi==23.1.0
pytest==7.4.3
pytest-django==4.7.0
